+++ app/components/GetSinglePageQuery.tsx
